refactor(gateway): split messenger_handler into a directory module (<1000)
messenger_handler.rs was 1499 lines. Convert it to messenger_handler/ with
focused submodules so each file is under 1000 lines:

- builders.rs: messenger construction from config (create_messenger + the
  per-platform build_* helpers).
- prompt.rs: per-message system-prompt assembly.
- media.rs: image attachment download/sniffing/caching (+ the not-yet-wired
  multimodal payload builder, kept behind #[allow(dead_code)]).

mod.rs (654 lines) keeps the polling loop, message dispatch, and the tool
loop in process_incoming_message. No behavior change.
