+++ crates/rustyclaw-gateway/src/messenger_handler/media.rs
@@ -1,0 +1,318 @@
+//! Image attachment handling for the messenger loop.
+//!
+//! Downloads / loads image attachments, sniffs their MIME type, caches them to
+//! disk, and (for the not-yet-wired vision path) assembles multimodal request
+//! payloads.
+
+use anyhow::{Context, Result};
+use serde_json::{Value, json};
+use tracing::{debug, trace};
+
+use rustyclaw_core::gateway::MediaRef;
+use rustyclaw_core::messengers::MediaAttachment;
+
+/// Maximum image size to download (10 MB).
+const MAX_IMAGE_SIZE: usize = 10 * 1024 * 1024;
+
+/// Supported image MIME types for vision models.
+const SUPPORTED_IMAGE_TYPES: &[&str] = &["image/jpeg", "image/png", "image/gif", "image/webp"];
+
+// ── Image Handling ──────────────────────────────────────────────────────────
+
+/// Downloaded image data ready for inclusion in model request.
+#[derive(Debug, Clone)]
+pub(crate) struct ImageData {
+    data: Vec<u8>,
+    #[allow(dead_code)]
+    mime_type: String,
+    pub(crate) media_ref: MediaRef,
+}
+
+/// Download an image from a URL.
+/// Download an image from a URL and cache locally.
+async fn download_image(
+    http: &reqwest::Client,
+    url: &str,
+    filename: Option<&str>,
+    cache_dir: &std::path::Path,
+) -> Result<ImageData> {
+    let response = http
+        .get(url)
+        .send()
+        .await
+        .context("Failed to fetch image")?;
+
+    // Check content type
+    let content_type = response
+        .headers()
+        .get("content-type")
+        .and_then(|v| v.to_str().ok())
+        .unwrap_or("application/octet-stream")
+        .split(';')
+        .next()
+        .unwrap_or("application/octet-stream")
+        .to_string();
+
+    if !SUPPORTED_IMAGE_TYPES.contains(&content_type.as_str()) {
+        anyhow::bail!("Unsupported image type: {}", content_type);
+    }
+
+    // Check content length if provided
+    if let Some(len) = response.content_length() {
+        if len as usize > MAX_IMAGE_SIZE {
+            anyhow::bail!("Image too large: {} bytes (max {})", len, MAX_IMAGE_SIZE);
+        }
+    }
+
+    let bytes = response.bytes().await.context("Failed to read image")?;
+
+    if bytes.len() > MAX_IMAGE_SIZE {
+        anyhow::bail!(
+            "Image too large: {} bytes (max {})",
+            bytes.len(),
+            MAX_IMAGE_SIZE
+        );
+    }
+
+    // Build media ref
+    let mut media_ref = MediaRef::new(content_type.clone());
+    media_ref.filename = filename.map(String::from);
+    media_ref.size = Some(bytes.len());
+    media_ref.url = Some(url.to_string());
+
+    // Cache to disk
+    let ext = mime_to_extension(&content_type);
+    let cache_path = cache_dir.join(format!("{}.{}", media_ref.id, ext));
+
+    if let Err(e) = tokio::fs::write(&cache_path, &bytes).await {
+        debug!(error = %e, path = %cache_path.display(), "Failed to cache image");
+    } else {
+        media_ref.local_path = Some(cache_path.to_string_lossy().to_string());
+    }
+
+    Ok(ImageData {
+        data: bytes.to_vec(),
+        mime_type: content_type,
+        media_ref,
+    })
+}
+
+/// Load an image from a local file path.
+async fn load_image_from_path(path: &str, cache_dir: &std::path::Path) -> Result<ImageData> {
+    use tokio::fs;
+
+    let data = fs::read(path).await.context("Failed to read image file")?;
+
+    if data.len() > MAX_IMAGE_SIZE {
+        anyhow::bail!(
+            "Image too large: {} bytes (max {})",
+            data.len(),
+            MAX_IMAGE_SIZE
+        );
+    }
+
+    // Detect MIME type from extension or magic bytes
+    let mime_type = detect_image_mime_type(path, &data)?;
+
+    // Build media ref
+    let mut media_ref = MediaRef::new(mime_type.clone());
+    media_ref.filename = std::path::Path::new(path)
+        .file_name()
+        .and_then(|n| n.to_str())
+        .map(String::from);
+    media_ref.size = Some(data.len());
+
+    // Copy to cache dir
+    let ext = mime_to_extension(&mime_type);
+    let cache_path = cache_dir.join(format!("{}.{}", media_ref.id, ext));
+
+    if let Err(e) = tokio::fs::write(&cache_path, &data).await {
+        debug!(error = %e, path = %cache_path.display(), "Failed to cache image");
+    } else {
+        media_ref.local_path = Some(cache_path.to_string_lossy().to_string());
+    }
+
+    Ok(ImageData {
+        data,
+        mime_type,
+        media_ref,
+    })
+}
+
+/// Get file extension for MIME type.
+fn mime_to_extension(mime: &str) -> &'static str {
+    match mime {
+        "image/jpeg" => "jpg",
+        "image/png" => "png",
+        "image/gif" => "gif",
+        "image/webp" => "webp",
+        _ => "bin",
+    }
+}
+
+/// Detect image MIME type from path extension or magic bytes.
+fn detect_image_mime_type(path: &str, data: &[u8]) -> Result<String> {
+    // Try extension first
+    let ext = std::path::Path::new(path)
+        .extension()
+        .and_then(|e| e.to_str())
+        .map(|e| e.to_lowercase());
+
+    if let Some(ext) = ext {
+        let mime = match ext.as_str() {
+            "jpg" | "jpeg" => "image/jpeg",
+            "png" => "image/png",
+            "gif" => "image/gif",
+            "webp" => "image/webp",
+            _ => return detect_from_magic_bytes(data),
+        };
+        return Ok(mime.to_string());
+    }
+
+    detect_from_magic_bytes(data)
+}
+
+/// Detect image type from magic bytes.
+fn detect_from_magic_bytes(data: &[u8]) -> Result<String> {
+    if data.len() < 4 {
+        anyhow::bail!("Data too small to detect image type");
+    }
+
+    // JPEG: FF D8 FF
+    if data.starts_with(&[0xFF, 0xD8, 0xFF]) {
+        return Ok("image/jpeg".to_string());
+    }
+
+    // PNG: 89 50 4E 47
+    if data.starts_with(&[0x89, 0x50, 0x4E, 0x47]) {
+        return Ok("image/png".to_string());
+    }
+
+    // GIF: GIF87a or GIF89a
+    if data.starts_with(b"GIF87a") || data.starts_with(b"GIF89a") {
+        return Ok("image/gif".to_string());
+    }
+
+    // WebP: RIFF....WEBP
+    if data.len() >= 12 && data.starts_with(b"RIFF") && &data[8..12] == b"WEBP" {
+        return Ok("image/webp".to_string());
+    }
+
+    anyhow::bail!("Could not detect image type from magic bytes")
+}
+
+/// Process media attachments and return image data.
+pub(crate) async fn process_attachments(
+    http: &reqwest::Client,
+    attachments: &[MediaAttachment],
+    cache_dir: &std::path::Path,
+) -> Vec<ImageData> {
+    // Ensure cache directory exists
+    if let Err(e) = tokio::fs::create_dir_all(cache_dir).await {
+        debug!(error = %e, path = %cache_dir.display(), "Failed to create cache dir");
+    }
+
+    let mut images = Vec::new();
+
+    for attachment in attachments {
+        // Skip non-image attachments
+        if let Some(mime) = &attachment.mime_type {
+            if !SUPPORTED_IMAGE_TYPES.contains(&mime.as_str()) {
+                continue;
+            }
+        }
+
+        // Try URL first, then path
+        let result = if let Some(url) = &attachment.url {
+            download_image(http, url, attachment.filename.as_deref(), cache_dir).await
+        } else if let Some(path) = &attachment.path {
+            load_image_from_path(path, cache_dir).await
+        } else {
+            continue;
+        };
+
+        match result {
+            Ok(img) => {
+                trace!(
+                    filename = %attachment.filename.as_deref().unwrap_or("unknown"),
+                    size_bytes = img.data.len(),
+                    media_id = %img.media_ref.id,
+                    "Downloaded image"
+                );
+                images.push(img);
+            }
+            Err(e) => {
+                debug!(error = %e, "Failed to process attachment");
+            }
+        }
+    }
+
+    images
+}
+
+/// Build a multi-modal user message with text and images.
+///
+/// For OpenAI-compatible APIs, this returns a content array:
+/// ```json
+/// {
+///   "role": "user",
+///   "content": [
+///     { "type": "text", "text": "What's in this image?" },
+///     { "type": "image_url", "image_url": { "url": "data:image/jpeg;base64,..." } }
+///   ]
+/// }
+/// ```
+#[allow(dead_code)]
+fn build_multimodal_user_message(text: &str, images: &[ImageData], provider: &str) -> Value {
+    use base64::{Engine, engine::general_purpose::STANDARD};
+
+    if images.is_empty() {
+        // Simple text message
+        return json!({
+            "role": "user",
+            "content": text
+        });
+    }
+
+    // Build content array with text and images
+    let mut content = Vec::new();
+
+    // Add text part
+    if !text.is_empty() {
+        content.push(json!({
+            "type": "text",
+            "text": text
+        }));
+    }
+
+    // Add image parts
+    for img in images {
+        let b64 = STANDARD.encode(&img.data);
+        let data_url = format!("data:{};base64,{}", img.mime_type, b64);
+
+        if provider == "anthropic" {
+            // Anthropic uses different format
+            content.push(json!({
+                "type": "image",
+                "source": {
+                    "type": "base64",
+                    "media_type": img.mime_type,
+                    "data": b64
+                }
+            }));
+        } else {
+            // OpenAI format (also works with many compatible APIs)
+            content.push(json!({
+                "type": "image_url",
+                "image_url": {
+                    "url": data_url
+                }
+            }));
+        }
+    }
+
+    json!({
+        "role": "user",
+        "content": content
+    })
+}
